I fixed some bugs and added the ability to transfer the current balance to another currency

+++ setup.sql
@@ -1,4 +1,4 @@
 create table users (
    id       serial primary key,
-   balance  integer
+   balance  numeric(15, 2)
 );
